suggested changes

+++ samples/test/support/management-api/deleteUser.ts
@@ -2,7 +2,11 @@
 import a18nClient from './a18nClient';
 
 export default async function(): Promise<void> {
-  await this.user.deactivate();
-  await this.user.delete();
-  await a18nClient.deleteProfile(this.a18nProfile.profileId);
+  if (this.user) {
+    await this.user.deactivate();
+    await this.user.delete();
+  }
+  if (this.a18nProfile) {
+    await a18nClient.deleteProfile(this.a18nProfile.profileId);
+  }
 }
